Create Day 3
The induction program on 4th August 2025 for students of Robotics & Al, ME, CE, EE, and ECE was a well-organized and insightful experience.
The day began with an energetic session by the Causmic Club from 8:30 to 9:30 AM in the auditorium, followed by an informative lecture by Dr.
Randhir Singh on the "Conduct of Examinations.
Later, an expert lecture on "Universal Human Values-2" was delivered by Dr. Priya Darshini, offering a deeper understanding of moral and ethical aspects of life. Main thing is we explore instill fundamental human values, promoting ethical behavior and harmonious living. This lecture cover topics like love, compassion, non-violence, righteousness, peace, service, renunciation, and constitutional values, encouraging individuals to understand and apply these principles in their lives.
The goal is to foster a holistic understanding of human values and their implications for individual conduct and relationships with others and nature