Day 5
The fifth day of our induction program at GNE
College was filled with energy, inspiration, and valuable insights. The day began with an informative lecture on the importance of fitness for engineers, highlighting how physical health plays a crucial role in maintaining mental sharpness and overall productivity in such a demanding profession. This was followed by an impactful session led by Brigadier P.S. Chima, who guided us through the structure and spirit of the National Cadet Corps (NCC). He spoke about the values of discipline, leadership, and service that the NCC instills in young individuals, and also shed light on the broader role of the National Defence Services. As someone who has attempted the NDA exam, I found this session particularly motivating. Although I couldn't complete the process at the time, the experience taught me valuable lessons and strengthened my desire to pursue opportunities in defence in the future. The day concluded with a thoughtful lecture by the Dean of our college, focusing on student welfare. It was a reassuring session that introduced us to the support systems available at GNE, showing that the college genuinely cares about our academic, emotional, and personal well-being. Overall, the day left a lasting impression & le, pushing me to strive harder and remindina me that every success
He spoke about the values of discipline, leadership, and service that the NCC instills in young individuals, and also shed light on the broader role of the National Defence Services. As someone who has attempted the NDA exam, I found this session particularly motivating. Although I couldn't complete the process at the time, the experience taught me valuable lessons and strengthened my desire to pursue opportunities in defence in the future. The day concluded with a thoughtful lecture by the Dean of our college, focusing on student welfare. It was a reassuring session that introduced us to the support systems available at GNE, showing that the college genuinely cares about our academic, emotional, and personal well-being. Overall, the day left a lasting impression on me, pushing me to strive harder and reminding me that every step-success or not-is a part of the journey.